+++ cincan/frontend.py
@@ -437,19 +437,7 @@ def main():
 
     mani_parser = subparsers.add_parser('manifest')
     image_default_args(mani_parser)
-
-    fanin_parser = subparsers.add_parser('fanin', help='Show fan-in to the given file')
-    fanin_parser.add_argument('file', help="File to analyze")
-    fanin_parser.add_argument('-d', '--max-depth', default=3, help='Maximum tree depth')
-    fanout_parser = subparsers.add_parser('fanout', help='Show fan-out from the given file')
-    fanout_parser.add_argument('file', help="File to analyze")
-    fanout_parser.add_argument('-d', '--max-depth', default=3, help='Maximum tree depth')
-
     help_parser = subparsers.add_parser('help')
-
-    commit_parser = subparsers.add_parser('commit')
-    image_default_args(commit_parser)
-
     if len(sys.argv) > 1:
         args = m_parser.parse_args(args=sys.argv[1:])
     else:
@@ -515,14 +503,6 @@ def main():
         if log.stderr:
             sys.stderr.buffer.write(log.stderr)
         sys.exit(log.exit_code)  # exit code
-    elif sub_command in {'fanin', 'fanout'}:
-        inspector = CommandInspector(CommandLogIndex(), pathlib.Path().resolve())
-        depth = int(args.max_depth)
-        if sub_command == 'fanout':
-            res = inspector.fanout(pathlib.Path(args.file).resolve(), depth)
-        else:
-            res = inspector.fanin(pathlib.Path(args.file).resolve(), depth)
-        print(res)
     elif sub_command == 'manifest':
         # sub command 'manifest'
         if len(args.tool) == 0:
@@ -535,28 +515,5 @@ def main():
         print(json.dumps(info, indent=2))
     elif sub_command == 'list':
         list_handler(args)
-    elif sub_command == 'commit':
-
-        log_path = str(pathlib.Path.home() / '.cincan/shared')
-
-        # change working dir where logs lie.
-        os.chdir(log_path)
-
-        print("check if git exists in current directory")
-
-        if os.path.exists('.git'):
-            print("if git exists, pull repo")
-            subprocess.call(["git", "pull"])
-            print("Add, commit and push logs")
-            subprocess.call(["git", "add", "."])
-            subprocess.call(["git", "status"])
-            subprocess.call(["git", "commit", "-m", "added log files from shared folder with cincan commit -command"])
-            subprocess.call(["git", "push"])
-        else:
-            print("Git doesn't exist. If you want to share your logs: Go to .cincan/shared folder")
-            print("type 'git init' and attach folder to remote repository for sharing logs")
-            print("git remote add origin git@gitlab.com:CinCan/log-sharing.git")
-            print("git pull origin master")
-            print("git branch --set-upstream-to=origin/master master")
     else:
         sys.exit(f"Unexpected sub command '{sub_command}")